Reverting Andreas changes again. I am really sorry. We should coordinate this better.
However checking in auto generated files is simply wrong. It creates an endless stream of pain. The first example is the problem with the revision numbers, that is only one permutation of the underlying problem that source and generated file will never be synchronous. The frequent changes will furthermore spam the history and bloat the repo.

In my opinion the right solution is to not 'svn export' the source files, but to work with a build type of directory. The script should do

svn up
perl makedoku.pl html
cp Homepage/* labvisa.de/

This is analogous to building other things like programs, tex documents etc.


git-svn-id: http://daimlerstr.de/svn/lab/trunk@630 d7904494-db04-0410-88ca-e47d1dcc2d9a