Created new POD file Lab::VISA::Installation and moved installation instructions from various places to there. Added this file to the documentation.
I am deleting installation.txt so that future modification will happen in the new POD. I admit that this is a bit of pain, due to all the lists. I have checked that all info (except the ones on doku building) are included.

If this deletion bothers anyone, just revert it.

git-svn-id: http://daimlerstr.de/svn/lab/trunk@652 d7904494-db04-0410-88ca-e47d1dcc2d9a