Revert "Minor Bug in TRMC support. Allowed connections were NONE."
This reverts commit a4bedc32e257bd80be89fe57b3dbe74549431776.

TRMC does really not use any connections. Just instantiate the instrument without connection parameters.
Communication with the LabVIEW control process is via local files.
(Maybe we should put this in the docs somewhere.)

(It could do it for non-standard setups, like the LabVIEW control program listening on a serial port,
but that is not coded yet anyway.)