Use Carp module for all warnings.
This will report all errors in the context of the measurement script.
To get a stack trace of an error, run 'perl -MCarp=verbose script.pl'.

The array @CARP_NOT, which holds a list of our packages, is auto generated by
Build.PL.

Carp seems to be in conflict with Hook::LexWrap (required for the caching in
Lab::Instrument). We provide a patched version in Lab::LexWrap.