The previous commit seams to break the Yoko.
The issue needs further discussion and testing.

Revert "Add "interface clear" to LinuxGPIB new connection initialization. When a NI-USB-HS is first plugged in, the ibdev call works, but subsequent read/write fails on the first try, subsequent tries work. (presumably the interface is cleared during the connection shutdown). Perhaps this is a change that would be better in the LinuxGpib package, but adding a 'ibclr' when starting a new connection should be mostly harmless, and puts the interface into a known state."

This reverts commit 69959438a7a7121f38addd02e7210df61895d207.